rk3399 - upgrade cv to 4.1.2

+++ install-opencv.sh
@@ -17,7 +17,7 @@ fi
 . /etc/friendlyelec-release
 
 if [ -d /usr/local/Trolltech/Qt-5.10.0-rk64one-sdk ]; then
-    CVSH=OpenCV-4.1.0-For-FriendlyELEC-RK3399.sh
+    CVSH=OpenCV-4.1.2-For-FriendlyELEC-RK3399.sh
     PyVER=3.6
 elif [ -d /usr/local/Trolltech/Qt-5.10.0-nexell32-sdk ]; then
     CVSH=OpenCV-3.4.2-For-FriendlyCore-S5Pxx18-armhf.sh
@@ -42,7 +42,7 @@ TOPPATH=$PWD
 mkdir -p .cache/download
 if [ ! -f .cache/${CVSH} ]; then
 	apt-get -y install curl
-        curl -o .cache/download/${CVSH} http://112.124.9.243/opencv/${CVSH}
+    curl -o .cache/download/${CVSH} http://112.124.9.243/opencv/${CVSH}
 	rc=$?; if [ $rc != 0 ]; then exit $rc; fi;
 	curl -o .cache/download/${CVSH}.hash.md5 http://112.124.9.243/opencv/${CVSH}.hash.md5
 	rc=$?; if [ $rc != 0 ]; then exit $rc; fi;
@@ -59,8 +59,8 @@ cd ${TOPPATH}
 
 # remove old packages
 apt-get -y purge libopencv*
-apt-get -y purge python-numpy
-apt-get -y autoremove
+apt-get -y purge python3-numpy python3-matplotlib python-numpy python-matplotlib
+apt-get autoremove
 
 # compiler 
 apt-get -y install build-essential
@@ -82,20 +82,14 @@ apt -y install mesa-common-dev libglu1-mesa-dev freeglut3-dev
 
 # python3
 apt-get -y install python3-dev python3-pip python3-tk
-pip3 install virtualenv virtualenvwrapper -i https://pypi.douban.com/simple
+pip3 install virtualenv virtualenvwrapper -i https://mirrors.ustc.edu.cn/pypi/web/simple
 
 # fix v4l2 header file
 (cd /usr/include/linux; [ -f videodev.h ] || ln -s ../libv4l1-videodev.h videodev.h)
 
-export VIRTUALENVWRAPPER_PYTHON=/usr/bin/python3
-export WORKON_HOME=$HOME/.virtualenvs
-export VIRTUALENVWRAPPER_HOOK_DIR=
-export ZSH_VERSION=
-source /usr/local/bin/virtualenvwrapper.sh
-mkvirtualenv cv
-
-pip install numpy -i https://pypi.douban.com/simple
-pip install matplotlib -i https://pypi.douban.com/simple
+# install numpy/matplotlib
+pip3 install numpy -i https://mirrors.ustc.edu.cn/pypi/web/simple
+pip3 install matplotlib -i https://mirrors.ustc.edu.cn/pypi/web/simple
 
 # install opencv
 chmod 755 ./.cache/${CVSH}
@@ -104,17 +98,39 @@ chmod 755 ./.cache/${CVSH}
 # ----------------------------
 # mk link
 
-cd ~/.virtualenvs/cv/lib/python${PyVER}/site-packages/
-rc=$?; if [ $rc != 0 ]; then exit $rc; fi;
-echo "enter ~/.virtualenvs/cv/lib/python${PyVER}/site-packages/, result: $rc"
-rm -f cv2.so
-ln -s /usr/local/lib/python${PyVER}/site-packages/cv2/python-${PyVER}/cv2.cpython-*.so cv2.so
-
 cd /usr/local/lib/python${PyVER}/site-packages/
 rc=$?; if [ $rc != 0 ]; then exit $rc; fi;
 echo "enter /usr/local/lib/python${PyVER}/site-packages/, result: $rc"
 rm -f cv2.so
 ln -s /usr/local/lib/python${PyVER}/site-packages/cv2/python-${PyVER}/cv2.cpython-*.so cv2.so
+
+# ---------------------------
+# mk virtualenv for pi user
+
+if [ -d /home/pi ]; then
+cat > /tmp/mkcv_for_piuser << EOL
+#!/bin/bash
+export VIRTUALENVWRAPPER_PYTHON=/usr/bin/python3
+export WORKON_HOME=\$HOME/.virtualenvs
+export VIRTUALENVWRAPPER_HOOK_DIR=
+export ZSH_VERSION=
+source /usr/local/bin/virtualenvwrapper.sh
+mkvirtualenv cv
+
+cd ~/.virtualenvs/cv/lib/python${PyVER}/site-packages/
+rc=\$?; if [ \$rc != 0 ]; then exit \$rc; fi;
+echo "enter ~/.virtualenvs/cv/lib/python${PyVER}/site-packages/, result: \$rc"
+rm -f cv2.so
+ln -s /usr/local/lib/python${PyVER}/site-packages/cv2/python-${PyVER}/cv2.cpython-*.so cv2.so
+EOL
+    chmod 755 /tmp/mkcv_for_piuser
+    su - pi -c "/tmp/mkcv_for_piuser"
+fi
+
+# ---------------------------
+# mk virtualenv for root user
+su - root -c "/tmp/mkcv_for_piuser"
+
 
 QtEnvScript=setqt5env-eglfs
 if [ x"${LINUXFAMILY}" = "xnanopi4" ]; then
@@ -135,9 +151,9 @@ echo "-   Testing OpenCV Installation  -"
 echo "----------------------------------"                                                                        
 echo ""                                                                                                          
 echo "Test python code:" 
-echo "# cd examples"                                                                                         
+echo "# cd examples/py"                                                                                         
 echo "# . cv-env.sh" 
-echo "# python py/ver.py" 
+echo "# python3 ver.py" 
 echo ""                                                                                        
 echo "-------" 
 echo "Test Qt5/C++ code:"
